requirements.txt
streamlit
duckdb
pandas
openpyxl
pydantic
pytest
google-generativeai
python-dotenv
